requirements.txt
alembic==1.12.0
annotated-types==0.6.0
anyio==3.7.1
click==8.1.7
colorama==0.4.6
et-xmlfile==1.1.0
fastapi==0.104.0
greenlet==3.0.0
h11==0.14.0
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.3
numpy==1.26.1
openpyxl==3.1.2
pandas==2.1.1
pydantic==2.4.2
pydantic_core==2.10.1
pyodbc==5.0.1
pyperclip==1.8.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.22
starlette==0.27.0
typing_extensions==4.8.0
tzdata==2023.3
uvicorn==0.23.2
